+++ src/utils/platform_manager.py
@@ -1,8 +1,8 @@
 import math
 import random
-from pygame.math import Vector2, clamp
-from utils.time import Time
-from utils.stage import Stage
+from pygame.math import Vector2
+from utils.data.time import Time
+from utils.gui.stage import Stage
 from physics.colliders import CircleCollider, PolygonCollider
 from physics.collisionhandler import CollisionHandler
 from entities.player import Player
@@ -13,104 +13,171 @@ from entities.platform import Platform
 class PlatformManager:
     current_platforms: list[Platform] = []
 
-    # Generation variables
-    __y_range = Vector2(100, 400)
-    __x_range = Vector2(100, 600)
-    __size_range = Vector2(20, 120)
+    # Static platforms occasionally
+        # Distance between them increases the higher the player climbs
+    # Mostly dynamic platforms
+        # Sideways movement
+        # Spawn either left or right, move to the other side and get destroyed
 
-    __prev_pos = Vector2(0, 0)
-    __current_dist = 0
-    __current_size = 0
+    # Static Platform Variables
+    dist_to_next = 200
+    next_y = 0
+
+    # Dynamic Platform Variables
+    spawn_delay = 3
+    spawn_timer = 0
+
+    start_platform_color = (0, 200, 100)
+    static_platform_color = (200, 200, 50)
+    dynamic_platform_color = (50, 200, 200)
+
+# Control Methods
 
     @classmethod
     def begin(cls):
-        cls.__prev_pos = Vector2(Stage.WIDTH / 2, Stage.HEIGHT * 0.8)
-        circle = CircleCollider(cls.__prev_pos.copy(), 100)
-        start_platform = Platform(True, circle, (0, 200, 100))
+        start_platform_coll = CircleCollider(Vector2(Stage.WIDTH / 2, Stage.bottom() - 150), 100)
+        start_platform = Platform(True, start_platform_coll, cls.start_platform_color)
         cls.current_platforms.append(start_platform)
 
-        cls.__set_values()
-        while (cls.__can_generate()):
-            cls.generate()
-
-    @classmethod
-    def __can_generate(cls) -> bool:
-        return cls.__prev_pos.y + Stage.Offset.y + cls.__current_size >= cls.__current_dist
-
-    @classmethod
-    def __set_values(cls):
-        cls.__current_dist = random.randint(cls.__y_range.x, cls.__y_range.y)
-        cls.__current_size = random.randint(cls.__size_range.x, cls.__size_range.y)
-
-    @classmethod
-    def __get_pos(cls) -> Vector2:
-        dx = random.randint(cls.__x_range.x, cls.__x_range.y)
-        x = random.randint(cls.__prev_pos.x - dx, cls.__prev_pos.x + dx)
-        x = clamp(x, 0, Stage.WIDTH)
-        return Vector2(x, cls.__prev_pos.y)
-
-    @classmethod
-    def generate(cls):
-        cls.destroy()
-
-        if (not cls.__can_generate()):
-            return
-        
-        cls.__prev_pos.y -= cls.__current_dist
-
-        if (random.random() < 0.5):
-            cls.__create_circle()
-        else:
-            cls.__create_polygon()
-        
-        cls.__prev_pos.x = cls.current_platforms[-1].coll.pos.x
-        cls.__set_values()        
-
-    @classmethod
-    def destroy(cls):
-        if (cls.current_platforms[0].coll.pos.y + Stage.Offset.y > Stage.HEIGHT + cls.__size_range.y):
-            cls.current_platforms.pop(0)
-
-    @classmethod
-    def __create_circle(cls):
-        circle = CircleCollider(cls.__get_pos(), cls.__current_size)
-        cls.current_platforms.append(Platform(True, circle, (255, 255, 255)))
-
-    @classmethod
-    def __create_polygon(cls):
-        vs = []
-        n = random.randint(3, 8)
-        for i in range(n):
-            v_norm = Vector2(math.cos(math.pi * 2 / n * i), math.sin(math.pi * 2 / n * i))
-            vs.append(v_norm * cls.__current_size)
-        poly = PolygonCollider(cls.__get_pos(), vs)
-        cls.current_platforms.append(Platform(True, poly, (255, 255, 255)))
+        cls.next_y = start_platform_coll.bounds.top - cls.dist_to_next
 
     @classmethod
     def update(cls):
-        for c in cls.current_platforms:
-            if (c.is_static):
-                cls.__static_check(c)
-            else:
-                c.move()
-                cls.__dynamic_check(c)
-
-    # Check player collision against a non moving platform
-    @classmethod
-    def __static_check(cls, c: Platform):
-        info = CollisionHandler.circle_collision(Player.instance.coll, c.coll)
-        if (info is not None):
-            Player.instance.controller.add_to_buffer(info)
-
-    # Check player collision against a moving platform
-    @classmethod
-    def __dynamic_check(cls, c: Platform):
-        info = CollisionHandler.circle_collision(Player.instance.coll, c.coll)
-        if (info is not None):
-            info.inherited_offset = c.vel * Time.dt
-            Player.instance.controller.add_to_buffer(info)
+        cls.__generate()
+        cls.__unload()
+        cls.__update_platforms()
 
     @classmethod
-    def render(cls):
+    def reset(cls):
+        cls.dist_to_next = 200
+        cls.spawn_timer = 0
+        cls.current_platforms.clear()
+
+    @classmethod
+    def draw(cls):
         for c in cls.current_platforms:
             c.draw()
+
+# Generation Methods
+
+    @classmethod
+    def __spawn_next_static(cls) -> bool:
+        return Stage.top() < cls.next_y
+    
+    @classmethod
+    def __spawn_next_dynamic(cls) -> bool:
+        return cls.spawn_timer > cls.spawn_delay
+    
+    @classmethod
+    def __increase_difficulty(cls):
+        cls.dist_to_next += 100
+
+    @classmethod
+    def __generate(cls):
+        while (cls.__spawn_next_static()):
+            cls.__create_static()
+            cls.__increase_difficulty()
+
+        cls.spawn_timer += Time.dt
+        if (cls.__spawn_next_dynamic()):
+            cls.spawn_timer = 0
+            cls.__create_dynamic()
+    
+    @classmethod
+    def __create_static(cls):
+        if (random.random() < 0.5):
+            r = random.randint(25, 150)
+            x = random.randint(r, Stage.WIDTH - r)
+            coll = CircleCollider(Vector2(x, cls.next_y - r), r)
+        else:
+            vs = []
+            n = random.randint(3, 8)
+            size = random.randint(25, 150)
+            offset_angle = random.random() * math.pi / 2
+            for i in range(n):
+                angle = math.pi * 2 / n * i + offset_angle
+                v_norm = Vector2(math.cos(angle), math.sin(angle))
+                vs.append(v_norm * size)
+            coll = PolygonCollider(Vector2(0, 0), vs)
+            x = random.randint(-coll.bounds.left, Stage.WIDTH - coll.bounds.right)
+            coll.pos = Vector2(x, cls.next_y - coll.bounds.bottom)
+        
+        platform = Platform(True, coll, cls.static_platform_color)
+        cls.current_platforms.append(platform)
+        cls.next_y = coll.bounds.top - cls.dist_to_next
+
+    @classmethod
+    def __create_dynamic(cls):
+        x_vel = random.randint(25, 150)
+
+        if (random.random() < 0.5):
+            r = random.randint(25, 150)
+            y = random.randint(Stage.top() + r, Stage.bottom() - r)
+            x = 0
+            if (random.random() < 0.5):
+                x = -r
+            else:
+                x = Stage.WIDTH + r
+                x_vel *= -1
+            coll = CircleCollider(Vector2(x, y), r)
+        else:
+            vs = []
+            n = random.randint(3, 8)
+            size = random.randint(25, 150)
+            offset_angle = random.random() * math.pi / 2
+            for i in range(n):
+                angle = math.pi * 2 / n * i + offset_angle
+                v_norm = Vector2(math.cos(angle), math.sin(angle))
+                vs.append(v_norm * size)
+            coll = PolygonCollider(Vector2(0, 0), vs)
+            y = random.randint(Stage.top() + coll.bounds.top, Stage.bottom() + coll.bounds.bottom)
+            x = 0
+            if (random.random() < 0.5):
+                x = -coll.bounds.right
+            else:
+                x = Stage.WIDTH - coll.bounds.left
+                x_vel *= -1
+            coll.pos = Vector2(x, y)
+        
+        platform = Platform(False, coll, cls.dynamic_platform_color)
+        platform.vel = Vector2(x_vel, 0)
+        cls.current_platforms.append(platform)
+
+    @classmethod
+    def __unload(cls):
+        n = len(cls.current_platforms)
+        i = 0
+        while (i < n):
+            if (cls.__unload_platform(i)):
+                cls.current_platforms.pop(i)
+                n -= 1
+                continue
+            i += 1
+    
+    @classmethod
+    def __unload_platform(cls, index: int) -> bool:
+        p = cls.current_platforms[index]
+
+        if (p.is_static):
+            if (p.coll.bounds.top > Stage.bottom()):
+                return True
+            return False
+        
+        if (p.vel.x > 0 and p.coll.bounds.left > Stage.WIDTH):
+            return True
+        if (p.vel.x < 0 and p.coll.bounds.right < 0):
+            return True
+        return False
+
+# Platform Updating
+
+    @classmethod
+    def __update_platforms(cls):
+        for c in cls.current_platforms:
+            if (not c.is_static):
+                c.move()
+            
+            info = CollisionHandler.circle_collision(Player.instance.coll, c.coll)
+            if (info is not None):
+                info.inherited_offset = Vector2(0, 0) if c.is_static else c.vel * Time.dt
+                Player.instance.controller.add_to_buffer(info)